fixed bucket storage

+++ src/s3_client.rs
@@ -72,6 +72,10 @@ impl S3Client {
 
     #[tracing::instrument(level = "info", skip(self))]
     pub async fn create_bucket(&self, _name: &str, region: &meta_store::BlobLocation) -> Result<(), s3s::S3Error> {
+        if self.cfg.storage.bucket.is_some() {
+            return Ok(())
+        }
+
         tracing::info!("creating bucket in the backing store");
         self.proxy
             .create_bucket(S3Request::new(try_!(CreateBucketInput::builder()
@@ -85,7 +89,10 @@ impl S3Client {
 
     #[tracing::instrument(level = "info", skip_all)]
     pub async fn delete_bucket(&self, _name: &str, region: &meta_store::BlobLocation) -> Result<(), s3s::S3Error> {
-        //try_!(self.client.create_bucket().set_bucket(Some(name.to_owned())).send().await);
+        if self.cfg.storage.bucket.is_some() {
+            return Ok(())
+        }
+        
         self.proxy
             .delete_bucket(S3Request::new(try_!(s3s::dto::builders::DeleteBucketInputBuilder::default()
                 .bucket(region.backend.clone())
